Fix: 이미지 없을 때 발생하는 버그 수정

+++ ANIS/src/main/java/com/npt/anis/ANIS/IdPhoto/service/IdPhotoServiceImpl.java
@@ -65,6 +65,10 @@ public class IdPhotoServiceImpl implements IdPhotoService {
 
     @Transactional
     public void updateImage(String studentId, MultipartFile file) throws IOException {
+        if (idPhotoRepository.findByName(studentId) == null) {
+            saveImage(studentId, file);
+            return;
+        }
         Path filePath = Paths.get(UPLOAD_DIR).resolve(studentId + ".png");
         Files.delete(filePath);
         Files.copy(file.getInputStream(), filePath);
